grc: Fixed an issue that was causing duplicate logging output
An additional line in grc/core/platform was causing duplicate log messages to be sent to the console.
This fixes the issue, so the logger can be set to DEBUG to catch all messages.
The console handler filters output to the console based on the user's selected level (defaults to WARNING).